// generated/src/app/api-models/stage-dnclm-dtl-sup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageDnclmDtlSup {
  batchId: string;
  transactionId: number;
  lineNo: number;
  subLineCode: string;
  claimStatus: string;
  processingStatus: string;
  seqProvId: number;
  seqMembId: number;
  adjudMethod: string;
  medDefCode: string;
  postDate: Date;
  companyCode: string;
  glRefCode: string;
  printFlag: string;
  printFlagChangeInd: string;
  eobId: string;
  raId: string;
  seqApTrans: number;
  seqAuthDtl: number;
  hiddenUserDef1: string;
  hiddenUserDef2: number;
  withholdSurplus: number;
  bmaAmt: number;
  bmaRsn: string;
  fullCvrgAmt: number;
  cobPatLiabCvrgAmt: number;
  anesInd: string;
  procCodeClass: string;
  seqCerulId: number;
  invalidFlags: string;
  geoZipRegion: string;
}